+++ pruebas/01-reading-list/fedekrenn/.eslintrc.cjs
@@ -17,6 +17,7 @@ module.exports = {
     'react-refresh/only-export-components': [
       'warn',
       { allowConstantExport: true }
-    ]
+    ],
+    'react/prop-types': 'off'
   }
 }
