+++ pruebas/01-reading-list/fedekrenn/vite.config.js
@@ -4,4 +4,8 @@ import react from '@vitejs/plugin-react-swc'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  test: {
+    globals: true,
+    environment: 'jsdom'
+  }
 })
